feat: implement /api/applications get method controller/service layer

+++ backend/src/main/resources/schema.sql
@@ -16,12 +16,12 @@ DROP TABLE IF EXISTS applications CASCADE;
 
 -- TABLES SET UP
 CREATE TABLE roles(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name VARCHAR(15)
 );
 
 CREATE TABLE accounts(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   registered_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -32,13 +32,13 @@ CREATE TABLE accounts(
 );
 
 CREATE TABLE accounts_roles_join(
-  account_id INT REFERENCES accounts(id),
-  role_id INT REFERENCES roles(id),
+  account_id BIGINT REFERENCES accounts(id),
+  role_id BIGINT REFERENCES roles(id),
   PRIMARY KEY (account_id, role_id)
 );
 
 CREATE TABLE pending_account_registrations(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   registered_at TIMESTAMP DEFAULT NOW() NOT NULL,
   first_name VARCHAR(255) NOT NULL,
   last_name VARCHAR(255) NOT NULL,
@@ -46,7 +46,7 @@ CREATE TABLE pending_account_registrations(
 );
 
 CREATE TABLE countries(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -54,9 +54,9 @@ CREATE TABLE countries(
 );
 
 CREATE TABLE universities(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
-  country_id INT REFERENCES countries(id),
+  country_id BIGINT REFERENCES countries(id),
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
   name VARCHAR(255) NOT NULL,
@@ -66,7 +66,7 @@ CREATE TABLE universities(
 );
 
 CREATE TABLE application_status(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -74,7 +74,7 @@ CREATE TABLE application_status(
 );
 
 CREATE TABLE interview_status(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -82,7 +82,7 @@ CREATE TABLE interview_status(
 );
 
 CREATE TABLE offer_status(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -90,7 +90,7 @@ CREATE TABLE offer_status(
 );
 
 CREATE TABLE response_status(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -98,7 +98,7 @@ CREATE TABLE response_status(
 );
 
 CREATE TABLE final_destination_status(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
@@ -106,20 +106,20 @@ CREATE TABLE final_destination_status(
 );
 
 CREATE TABLE applications(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   uuid UUID DEFAULT gen_random_uuid() UNIQUE,
-  account_id INT REFERENCES accounts(id),
+  account_id BIGINT REFERENCES accounts(id),
   created_at TIMESTAMP DEFAULT now() NOT NULL,
   last_updated_at TIMESTAMP DEFAULT now() NOT NULL,
-  country INT REFERENCES countries(id),
-  university INT REFERENCES universities(id),
+  country BIGINT REFERENCES countries(id),
+  university BIGINT REFERENCES universities(id),
   course_name VARCHAR(255) NOT NULL,
   minor_subject VARCHAR(255),
-  programme_length INT NOT NULL,
-  application_status_id INT REFERENCES application_status(id),
-  interview_status_id INT REFERENCES interview_status(id),
-  offer_status_id INT REFERENCES offer_status(id),
-  response_status_id INT REFERENCES response_status(id),
-  final_destination_status_id INT REFERENCES final_destination_status(id),
+  programme_length BIGINT NOT NULL,
+  application_status_id BIGINT REFERENCES application_status(id),
+  interview_status_id BIGINT REFERENCES interview_status(id),
+  offer_status_id BIGINT REFERENCES offer_status(id),
+  response_status_id BIGINT REFERENCES response_status(id),
+  final_destination_status_id BIGINT REFERENCES final_destination_status(id),
   notes VARCHAR(1500)
 );
